requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
affine=2.3.0=pyhd3eb1b0_0
aom=3.9.1=h7bae524_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=24.3.0=py312hca03da5_0
aws-c-auth=0.8.1=hfc2798a_0
aws-c-cal=0.8.1=hc8a0bd2_3
aws-c-common=0.10.6=h5505292_0
aws-c-compression=0.3.0=hc8a0bd2_5
aws-c-event-stream=0.5.0=h54f970a_11
aws-c-http=0.9.2=h96aa502_4
aws-c-io=0.15.3=haba67d1_6
aws-c-mqtt=0.11.0=h24f418c_12
aws-c-s3=0.7.9=hf37e03c_1
aws-c-sdkutils=0.2.2=hc8a0bd2_0
aws-checksums=0.2.2=hc8a0bd2_4
aws-crt-cpp=0.29.9=ha81f72f_2
aws-sdk-cpp=1.11.489=h0e5014b_0
azure-core-cpp=1.14.0=hd50102c_0
azure-identity-cpp=1.10.0=hc602bab_0
azure-storage-blobs-cpp=12.13.0=h7585a09_1
azure-storage-common-cpp=12.8.0=h9ca1f76_1
azure-storage-files-datalake-cpp=12.12.0=hcdd55da_1
black=24.10.0=py312hca03da5_0
blas=1.0=openblas
blosc=1.21.6=h7dd00d9_1
bokeh=3.6.2=py312hca03da5_0
bottleneck=1.4.2=py312ha86b861_0
branca=0.6.0=py312hca03da5_0
brotli-python=1.0.9=py312h313beb8_9
bzip2=1.0.8=h80987f9_6
c-ares=1.34.4=h5505292_0
ca-certificates=2025.2.25=hca03da5_0
cartopy=0.24.1=py312h313beb8_0
cdsapi=0.7.5=pyhd8ed1ab_1
certifi=2025.1.31=py312hca03da5_0
cf_xarray=0.10.2=pyhd8ed1ab_0
cffi=1.17.1=py312h3eb5a62_1
cfgrib=0.9.15.0=pyhd8ed1ab_0
cftime=1.6.4=py312h755e627_1
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=py312hca03da5_0
click-plugins=1.1.1=pyhd3eb1b0_0
cligj=0.7.2=pyhd3eb1b0_0
cloudpickle=3.0.0=py312hca03da5_0
colorama=0.4.6=py312hca03da5_0
contourpy=1.3.1=py312h48ca7d4_0
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=1.0.1=py312h80987f9_0
dask=2024.12.1=py312hca03da5_0
dask-core=2024.12.1=py312hca03da5_0
dask-expr=1.1.21=py312hca03da5_0
datapi=0.2.1=pyhd8ed1ab_0
dav1d=1.2.1=h80987f9_0
distributed=2024.12.1=py312hca03da5_0
eccodes=2.39.0=hd5c539c_1
esmf=8.7.0=nompi_hab24014_1
esmpy=8.7.0=pyhecae5ae_1
et_xmlfile=1.1.0=py312hca03da5_1
executing=0.8.3=pyhd3eb1b0_0
expat=2.6.4=h313beb8_0
findlibs=0.0.5=pyhd8ed1ab_0
folium=0.19.5=pyhd8ed1ab_0
fonttools=4.55.3=py312h80987f9_0
freetype=2.12.1=hadb7bae_2
freexl=2.0.0=ha3de405_0
fsspec=2024.12.0=py312hca03da5_0
geopandas=1.0.1=py312hca03da5_0
geopandas-base=1.0.1=py312hca03da5_0
geos=3.13.1=hc9a1286_0
geotiff=1.7.4=hbef4fa4_0
gflags=2.2.2=h313beb8_1
giflib=5.2.2=h80987f9_0
glog=0.7.1=heb240a5_0
hdf4=4.2.15=h2ee6834_7
hdf5=1.14.4=nompi_ha698983_105
heapdict=1.0.1=pyhd3eb1b0_0
icecream=2.1.4=pyhd8ed1ab_0
icu=75.1=hfee45f7_0
idna=3.7=py312hca03da5_0
jasper=4.2.4=h6c4e4ef_0
jinja2=3.1.5=py312hca03da5_0
joblib=1.4.2=py312hca03da5_0
json-c=0.18=he4178ee_0
kiwisolver=1.4.8=py312h313beb8_0
krb5=1.21.3=h237132a_0
lcms2=2.17=h7eeda09_0
lerc=4.0.0=h313beb8_0
libabseil=20240722.0=cxx17_h07bc746_4
libaec=1.1.3=hebf3989_0
libarchive=3.7.7=h3b16cec_3
libarrow=19.0.1=hf62c6bc_2_cpu
libarrow-acero=19.0.1=hf07054f_2_cpu
libarrow-dataset=19.0.1=hf07054f_2_cpu
libarrow-substrait=19.0.1=h4239455_2_cpu
libavif16=1.2.0=hf9d1e0e_0
libbrotlicommon=1.1.0=hd74edd7_2
libbrotlidec=1.1.0=hd74edd7_2
libbrotlienc=1.1.0=hd74edd7_2
libcrc32c=1.1.2=hc377ac9_0
libcurl=8.12.1=h73640d1_0
libcxx=19.1.7=ha82da77_0
libde265=1.0.15=h2ffa867_0
libdeflate=1.23=hec38601_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h93a5062_2
libevent=2.1.12=h02f6b3c_1
libexpat=2.6.4=h286801f_0
libffi=3.4.4=hca03da5_1
libgdal-core=3.10.2=h0528216_1
libgfortran=5.0.0=13_2_0_hd922786_3
libgfortran5=13.2.0=hf226fd6_3
libgoogle-cloud=2.36.0=hdbe95d5_0
libgoogle-cloud-storage=2.36.0=h7081f7f_0
libgrpc=1.67.1=h0a426d6_2
libheif=1.19.7=gpl_h79e6334_100
libiconv=1.18=hfe07756_1
libjpeg-turbo=3.0.0=hb547adb_1
libkml=1.3.0=he250239_1021
libllvm15=15.0.7=h4429f82_5
liblzma=5.6.4=h39f12f2_0
libnetcdf=4.9.2=nompi_h6569565_116
libnghttp2=1.64.0=h6d7220d_0
libopenblas=0.3.21=h269037a_0
libopentelemetry-cpp=1.18.0=h0c05b2d_1
libopentelemetry-cpp-headers=1.18.0=hce30654_1
libparquet=19.0.1=h636d7b7_2_cpu
libpng=1.6.47=h3783ad8_0
libprotobuf=5.28.3=h3bd63a1_1
libre2-11=2024.07.02=h07bc746_2
librttopo=1.1.0=h26cc057_18
libspatialite=5.1.0=haf574c3_13
libsqlite=3.49.1=h3f77e49_1
libssh2=1.11.1=h9cc3647_0
libthrift=0.21.0=h64651cc_0
libtiff=4.7.0=h551f018_3
libutf8proc=2.10.0=hda25de7_0
libwebp-base=1.5.0=h2471fea_0
libxcb=1.17.0=hdb1d25a_0
libxml2=2.13.6=h178c5d8_0
libzip=1.11.2=h1336266_0
libzlib=1.3.1=h8359307_2
llvm-openmp=14.0.6=hc6e5704_0
llvmlite=0.44.0=py312h313beb8_0
locket=1.0.0=py312hca03da5_0
lz4=4.3.3=py312hf263c89_2
lz4-c=1.10.0=h286801f_1
lzo=2.10=h1a28f6b_2
mapclassify=2.4.3=pyhd3eb1b0_0
markupsafe=3.0.2=py312h80987f9_0
matplotlib-base=3.10.0=py312h7ef442a_0
minizip=4.0.7=hff1a8ea_3
msgpack-python=1.0.3=py312h48ca7d4_0
multiurl=0.3.3=pyhd8ed1ab_1
mypy_extensions=1.0.0=py312hca03da5_0
ncurses=6.5=h5e97a16_3
netcdf-fortran=4.6.1=nompi_h1a59fa5_108
netcdf4=1.7.2=nompi_py312haae1a11_101
networkx=3.4.2=py312hca03da5_0
nlohmann_json=3.11.2=h313beb8_0
numba=0.61.0=py312h313beb8_1
numexpr=2.10.1=py312h5d9532f_0
numpy=2.1.3=py312h7f4fdc5_0
numpy-base=2.1.3=py312he047099_0
openjpeg=2.5.3=h8a3d83b_0
openpyxl=3.1.5=py312h80987f9_0
openssl=3.4.1=h81ee809_0
orc=2.1.1=h74f7c94_0
packaging=24.2=py312hca03da5_0
pandas=2.2.3=py312hcf29cfe_0
partd=1.4.2=py312hca03da5_0
pathspec=0.10.3=py312hca03da5_0
pcre2=10.44=h297a79d_2
pexpect=4.8.0=pyhd3eb1b0_3
pillow=11.1.0=py312h50aef2c_0
pip=25.0=py312hca03da5_0
platformdirs=3.10.0=py312hca03da5_0
proj=9.5.1=h1318a7e_0
prometheus-cpp=1.3.0=h0967b3e_0
psutil=5.9.0=py312h80987f9_1
pthread-stubs=0.3=h1a28f6b_1
ptyprocess=0.7.0=pyhd3eb1b0_2
pyarrow=19.0.1=py312h1f38498_0
pyarrow-core=19.0.1=py312hc40f475_0_cpu
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=py312hca03da5_1
pyogrio=0.10.0=py312hfd5e53c_1
pyparsing=3.2.0=py312hca03da5_0
pyproj=3.7.1=py312h4b98159_0
pyshp=2.3.1=py312hca03da5_0
pysocks=1.7.1=py312hca03da5_0
python=3.12.2=hdf0ec26_0_cpython
python-dateutil=2.9.0post0=py312hca03da5_2
python-eccodes=2.37.0=py312h755e627_0
python-lmdb=1.6.2=py312h313beb8_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.12=5_cp312
pytz=2024.1=py312hca03da5_0
pyyaml=6.0.2=py312h80987f9_0
rasterio=1.4.3=py312h02264c4_0
rav1e=0.6.6=h69fbcac_2
re2=2024.07.02=h6589ca4_2
readline=8.2=h1a28f6b_0
requests=2.32.3=py312hca03da5_1
rioxarray=0.18.1=py312hca03da5_0
scikit-learn=1.6.1=py312h313beb8_0
scipy=1.15.1=py312ha409365_0
seaborn=0.13.2=py312hca03da5_2
setuptools=72.1.0=py312hca03da5_0
shapely=2.0.7=py312hf733f26_1
six=1.16.0=pyhd3eb1b0_1
snappy=1.2.1=h98b9ce2_1
snuggs=1.4.7=pyhd3eb1b0_0
sortedcontainers=2.4.0=pyhd3eb1b0_0
sparse=0.15.4=py312hca03da5_0
sqlite=3.49.1=hd7222ec_1
svt-av1=3.0.0=h8ab69cd_0
tbb=2021.8.0=h48ca7d4_0
tblib=1.7.0=pyhd3eb1b0_0
threadpoolctl=3.5.0=py312h989b03a_0
tk=8.6.13=h5083fa2_1
toolz=1.0.0=py312hca03da5_0
tornado=6.4.2=py312h80987f9_0
tqdm=4.67.1=py312h989b03a_0
typing_extensions=4.12.2=py312hca03da5_0
tzdata=2025a=h04d1e81_0
unicodedata2=15.1.0=py312h80987f9_1
uriparser=0.9.8=h00cdb27_0
urllib3=2.3.0=py312hca03da5_0
wheel=0.45.1=py312hca03da5_0
x265=3.5=hbc6ce65_3
xarray=2024.11.0=py312hca03da5_0
xerces-c=3.2.5=h92fc2f4_2
xesmf=0.8.8=pyhd8ed1ab_1
xorg-libxau=1.0.12=h5505292_0
xorg-libxdmcp=1.1.5=hd74edd7_0
xyzservices=2022.9.0=py312hca03da5_1
xz=5.6.4=h80987f9_1
yaml=0.2.5=h1a28f6b_0
zict=3.0.0=py312hca03da5_0
zlib=1.3.1=h8359307_2
zstd=1.5.7=h6491c7d_1
